+++ CHANGELOG.md
@@ -29,6 +29,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Failing import on Windows
 - Subscription are now compared case sensitive (as it should have been)
+- `ValueClient` documentation for `prop_change_wait` timeout parameter
+- `ValueClient.prop_get` now uses `max_age` for `get` and correctly uses local
+  cache
 
 ### Removed
 - `rpc_login` and `rpc_login_url` in favor of `RpcLogin`